backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Caching>=2.1.0
Flask-Limiter>=4.1.0
Flask-JWT-Extended>=4.5.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.8.0
requests>=2.28.0
pydantic>=2.0.0
sentence-transformers==2.2.2
arxiv==1.4.4
numpy==1.26.4
huggingface-hub==0.20.3
pgvector==0.3.4
transformers==4.33.3
torch==2.2.1
accelerate==0.27.2
langchain==0.1.16
langchain-community==0.0.32

